Update to master HEAD and solve :
Conflicts:
	css/jstree/style.css
	css/palettes/aerialgreen.css
	css/palettes/auror.css
	css/palettes/automn.css
	css/palettes/classic.css
	css/palettes/clockworkorange.css
	css/palettes/dark.css
	css/palettes/flood.css
	css/palettes/greenflat.css
	css/palettes/hipster.css
	css/palettes/icecream.css
	css/palettes/lightblue.css
	css/palettes/premiumred.css
	css/palettes/teclib.css
	css/palettes/vintage.css
	css/print.css
	css/style_install.css
	css/styles_ie.css
	css/tab-scroller-menu.css
	front/document.form.php
	front/dropdown.common.form.php
	inc/event.class.php
	inc/rulecollection.class.php
	install/index.php
	install/update_031_04.php
	install/update_042_05.php
	install/update_04_042.php
	install/update_051_06.php
	install/update_05_051.php
	install/update_065_068.php
	install/update_0681_07.php
	install/update_068_0681.php
	install/update_06_065.php
	install/update_0712_0713.php
	install/update_0713_072.php
	install/update_071_0712.php
	install/update_0721_0722.php
	install/update_0722_0723.php
	install/update_0723_078.php
	install/update_072_0721.php
	install/update_0781_0782.php
	install/update_0782_080.php
	install/update_078_0781.php
	install/update_07_071.php
	install/update_0801_0803.php
	install/update_0803_083.php
	install/update_080_0801.php
	install/update_0831_0833.php
	install/update_0831_084.php
	install/update_083_0831.php
	install/update_0841_0843.php
	install/update_0843_0844.php
	install/update_0845_0846.php
	install/update_084_0841.php
	install/update_084_085.php
	install/update_085_0853.php
	install/update_to_031.php
	js/fileupload.js
	scripts/add_creation_date.php
	scripts/compute_dictionnary.php
	scripts/softcat_mass_compute.php
	tools/checkforms.php
	tools/checkfunction.php
	tools/cleanhistory.php
	tools/diff_plugin_locale.php
	tools/findtableswithoutclass.php
	tools/fk_generate.php
	tools/generate_bigdump.function.php
	tools/generate_bigdump.php
	tools/notincludedlanguages.php
	tools/phpunit/AllTests.php
	tools/phpunit/Framework/AllTests.php
	tools/phpunit/Framework/CommonDBTM/AllTests.php
	tools/phpunit/Framework/CommonDBTM/CanCheck.php
	tools/phpunit/Framework/CommonDBTM/DeleteRestore.php
	tools/phpunit/Framework/Dropdown/AllTests.php
	tools/phpunit/Framework/Dropdown/Import.php
	tools/phpunit/Framework/Dropdown/TreeCache.php
	tools/phpunit/Framework/Version.php
	tools/phpunit/Install/AllTests.php
	tools/phpunit/System/AllTests.php
	tools/test_allfiles.php
	tools/test_langfiles.php
	tools/testmail.php
	tools/testunit.php
	tools/update_registered_ids.php

+++ front/ticket.form.php
@@ -31,6 +31,8 @@
  */
 
 
+use Glpi\Event;
+
 include ('../inc/includes.php');
 
 Session::checkLoginUser();
@@ -155,14 +157,17 @@ if (isset($_POST["add"])) {
               sprintf(__('%s adds an actor'), $_SESSION["glpiname"]));
    Html::redirect($CFG_GLPI["root_doc"]."/front/ticket.form.php?id=".$_POST['tickets_id']);
 } else if (isset($_REQUEST['delete_document'])) {
-   $document_item = new Document_Item;
-   $found_document_items = $document_item->find("itemtype = 'Ticket' ".
-                                                " AND items_id = ".intval($_REQUEST['tickets_id']).
-                                                " AND documents_id = ".intval($_REQUEST['documents_id']));
-   foreach ($found_document_items  as $item) {
-      $document_item->delete($item, true);
+   $doc = new Document();
+   $doc->getFromDB(intval($_REQUEST['documents_id']));
+   if ($doc->can($doc->getID(), UPDATE)) {
+      $document_item = new Document_Item;
+      $found_document_items = $document_item->find("itemtype = 'Ticket' ".
+                                                   " AND items_id = ".intval($_REQUEST['tickets_id']).
+                                                   " AND documents_id = ".$doc->getID());
+      foreach ($found_document_items  as $item) {
+         $document_item->delete($item, true);
+      }
    }
-
    Html::back();
 }
 
@@ -199,7 +204,9 @@ if (isset($_GET["id"]) && ($_GET["id"] > 0)) {
    Html::header(__('New ticket'), '', "helpdesk", "ticket");
    unset($_REQUEST['id']);
    // Add a ticket from item : format data
-   if (isset($_REQUEST['_add_fromitem'])) {
+   if (isset($_REQUEST['_add_fromitem'])
+       && isset($_REQUEST['itemtype'])
+       && isset($_REQUEST['items_id'])) {
       $_REQUEST['items_id'] = array($_REQUEST['itemtype'] => array($_REQUEST['items_id']));
    }
    $track->display($_REQUEST);
